Reformat aad related codes
