+++ src/Microsoft.Azure.SignalR.Common/Auth/AccessKey.cs
