+++ src/Microsoft.Azure.SignalR.Common/Auth/MicrosoftEntra/MicrosoftEntraTokenProvider.cs
@@ -1,0 +1,19 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Threading.Tasks;
+
+namespace Microsoft.Azure.SignalR;
+
+internal class MicrosoftEntraTokenProvider : IAccessTokenProvider
+{
+    private readonly MicrosoftEntraAccessKey _accessKey;
+
+    public MicrosoftEntraTokenProvider(MicrosoftEntraAccessKey accessKey)
+    {
+        _accessKey = accessKey ?? throw new ArgumentNullException(nameof(accessKey));
+    }
+
+    public Task<string> ProvideAsync() => _accessKey.GetMicrosoftEntraTokenAsync();
+}
