+++ src/Microsoft.Azure.SignalR.Common/Endpoints/ServiceEndpoint.cs
@@ -86,7 +86,7 @@ namespace Microsoft.Azure.SignalR
                 {
                     lock (_lock)
                     {
-                        _accessKey ??= new AccessKeyForMicrosoftEntra(_serviceEndpoint, _tokenCredential, ServerEndpoint);
+                        _accessKey ??= new MicrosoftEntraAccessKey(_serviceEndpoint, _tokenCredential, ServerEndpoint);
                     }
                 }
                 return _accessKey;
